put variablles to dotenvValidator file
