+++ dotenvValidator.js
@@ -1,5 +1,36 @@
 const joi = require('@hapi/joi');
 
+const dotenvVariables = [
+  'MONGO_URL',
+  'SECRET',
+  'EXPIRES_IN',
+  'BASE_URI',
+  'TEST_BASE_URI',
+  'KEY_VAULTS_URI',
+  'IMAGE_LINK',
+  'STORAGE_ACCOUNT',
+  'ACCESS_KEY',
+  'AZURE_HOST',
+  'PAYMENT_SECRET',
+  'PAYMENT_MERCHANT_ID',
+  'NOVA_POSHTA_API_LINK',
+  'NOVA_POSHTA_API_KEY',
+  'PAYMENT_API_LINK',
+  'MAIL_USER',
+  'MAIL_PASS',
+  'MAIL_HOST',
+  'MAIL_PORT',
+  'RECOVERY_EXPIRE',
+  'FRONT_BASE_URI',
+  'TEST_BASE_URI',
+  'CURRENCY_API_URL',
+  'CONFIRMATION_SECRET',
+  'RECOVERY_EXPIRE',
+  'SUPER_ADMIN_EMAIL',
+  'SUPER_ADMIN_PASSWORD',
+  'SENDGRID_API_KEY',
+];
+
 function dotenvValidator(processEnv) {
   const envSchema = joi.object({
     MONGO_URL: joi
@@ -69,4 +100,4 @@ function dotenvValidator(processEnv) {
   return environment;
 }
 
-module.exports = dotenvValidator;
+module.exports = { dotenvValidator, dotenvVariables };
